[BACKPORT 2025.1][PLAT-17417] Enable and disable CP not working for dual NIC case
Summary:
Enable and disable CP not working for dual NIC case.

Earlier we had a fix in https://phorge.dev.yugabyte.com/D42741, where we find the correct ysql port to override (when the user passes in a custom value for the gflag `pgsql_proxy_bind_address`) based on the `enableConnectionPooling` flag in the universe details.

But we should not get the value of `enableConnectionPooling` from the universe, instead we should get it from the taskParams. The earlier case would only work when creating the universe, but if the universe is already created and then we want to enable/disable CP, the existing universe is not already updated at the time with the user required value of `enableConnectionPooling`. This value needs to be retrieved from the taskParams since that is the operation the user would like to perform now - which has the correct intended value and not the existing value.

Original commit: 784777729be81adbc369934ef1cf676647abfb25 / D43496

Test Plan:
Manually tested the following:
Created uni with CP enabled and `"pgsql_proxy_bind_address": "0.0.0.0:5433"`. Ensured tserver gflag value is `"pgsql_proxy_bind_address": "0.0.0.0:6433"`
Disabled CP. Ensured tserver gflag is `"pgsql_proxy_bind_address": "0.0.0.0:5433"` instead of `<IP>:5433`
Enabled CP. Ensured tserver gflag is `"pgsql_proxy_bind_address": "0.0.0.0:6433"` instead of `<IP>:6433`

Before this change, for dual NIC case, it was `<IP1>:6433,<IP2>:6433`. Now it is correct with value 0.0.0.0:6433`.

Reviewers: sneelakantan

Reviewed By: sneelakantan

Differential Revision: https://phorge.dev.yugabyte.com/D43515