[#26853] docdb: Track RBS progress and report it on master UI
Summary:
This diff adds information about RBS progress (bytes transferred, total bytes to transfer, and duration) to the TabletStatusListener. This info is aggregated across all tservers and shown on the master's Cluster Balancer page.

On the tserver, we now update the progress info in the TabletStatusListener after every SST chunk download (64 MiB by default), instead of only when downloading a new WAL segment (which wasn't a useful progress indicator because the SSTs account for most of the data in a tablet).

Screenshot: {F358682}

**Upgrade / Downgrade Safety**
The `GetActiveRbsInfo` RPC being added is new and only used on demand when the UI page is accessed.
Jira: DB-16262

Test Plan:
```
./yb_build.sh --cxx_test integration-tests_master_path_handlers-itest --gtest_filter MasterPathHandlersItest.ClusterBalancerOngoingRbs
./yb_build.sh --cxx-test remote_bootstrap_file_downloader-test
./yb_build.sh --cxx-test async_task_result_collector-test
```

Reviewers: zdrudi

Reviewed By: zdrudi

Subscribers: hsunder, ybase

Differential Revision: https://phorge.dev.yugabyte.com/D43534