[#26972] DocDB: Fix pg_locks failure in case of active advisory lock waiters
Summary:
When requesting a advisory lock, we seem to take weak locks on the prefix as we do with row level locks. But during the apply phase, we just seem to apply the strong intent corresponding to the full doc key.

As part of `pg_locks`,  we fetch both waiting and granted locks. When fetching waiting locks, we iterate over the intents for which conflict resolution is being run. This creates a problem for advisory lock waiters since the ysql backend expects the lock row to adhere to a specific format, and looks for fields like dboic, class/object oid etc.

This diff fixes the issue by returning just the row corresponding to the full doc key of the advisory lock waiter when serving `pg_locks` query.

TBD if weak prefix locking is required for advisory locks. It appears like we can proceed with strong locking the full doc key alone, since we don't deal with delete table etc on advisory lock table.
Jira: DB-16433

Test Plan: ./yb_build.sh --cxx-test pg_advisory_lock-test --gtest_filter PgAdvisoryLockTest.PgLocksWithWaiters

Reviewers: yyan, rthallam

Reviewed By: yyan, rthallam

Subscribers: yql, ybase

Differential Revision: https://phorge.dev.yugabyte.com/D43565