[#26965] DocDB: Support backup/restore for vector indexes
Summary:
To ensure the proper functionality of backup and restore operations for vector indexes, it is essential to maintain the correct file system path where the vector index data is stored.

The vector index data resides in a dedicated subdirectory within the table's data folder, following the naming convention:
`tablet-{tablet_id}.vi-{vector_index_id}`

For backup and restore procedures to work correctly, it is crucial that the `vector_index_id` remains consistent throughout the entire process. This identifier must not change between the backup being taken and the subsequent restore operation, as it is intrinsically tied to the physical storage location of the vector index data.

Following the changes introduced in commit e9aeaec84f71ffc365db674515e092a438a9707b/D43159, we now have a reliable mechanism to ensure this consistency. We can depend on the `relfilenode` identifier, which serves as a stable reference point. The key property of `relfilenode` is that it remains unchanged for a specific vector index even after the database has been restored, providing the necessary continuity for the backup/restore process to function correctly.

This stability in `relfilenode` allows the system to correctly locate and associate the vector index data with its corresponding index structure in the restored database environment, ensuring data integrity and operational correctness throughout the backup and restore lifecycle.

Also improved mini cluster test infrastructure to simplify usage of ybcontroller in mini cluster tests.
Jira: DB-16425

Test Plan: PgVectorIndexTest.Backup/*

Reviewers: zdrudi

Reviewed By: zdrudi

Subscribers: ybase, yql

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D43540