[BACKPORT 2025.1][PLAT-17387][YBA CLI]XCluster replication adds extra table to replication by itself
Summary:
CLI used `tables` in the xCluster config as the source of tables existing in the replication. This was incorrect since the list also contains tables that are extra on source/target etc. The fix here is to filter the tables IDs based on table details, and consider only the `Running` tables as part of edit operations.

Original commit: 9e44b4a8401b13e7e08051009e7112fd2fafbdfd / D43407

Test Plan: Manually remove tables that are extra ob target and check if the tables extra on source are not added in the replication

Reviewers: skurapati

Reviewed By: skurapati

Differential Revision: https://phorge.dev.yugabyte.com/D43576