[BACKPORT 2025.1][#25880] docdb: Use pgsupervisor to kill pg sessions when a tserver gets a new ysql lease.
Summary:
In this diff there's also conditional logic to start the pg postmaster. If the YSQL lease feature is enabled, the tablet server will not spawn a pg postmaster on boot and instead will wait until it gets a ysql lease to spawn a postmaster and begin accepting pg connections.

Added a new state, kPaused, to the process supervisor. The process supervisor now has 4 states:
kNotStarted, kRunning, kStopping, kPaused.

These states apply to the supervisor itself, not the supervised process. They control the
behaviour of the monitor thread.

kNotStarted - the ProcessSupervisor is in this state after construction until Start is called.
The monitor thread and the supervised process do not yet exist.
  possible state transitions:
      kRunning
      kStopping
kRunning - when Start is called, the ProcessSupervisor transitions to kRunning. It spawns the
process and creates a monitor thread. The monitor thread will respawn the supervised process if
it dies.
  possible state transitions:
      kPaused
      kStopping

kPaused - the supervised process is killed and the monitor thread will not respawn it.
  possible state transitions:
      kRunning
      kStopping

kStopping - the supervised process is sent signals to stop and the monitor thread should stop.
  possible state transitions:
      (none)
Jira: DB-15187

Original commit: 8a1646bcc30a670a8fdc2c1467e1bbd968fda9c3 / D43283

Test Plan:
```
./yb_build.sh fastdebug --cxx-test object_lock-test --gtest_filter 'ExternalObjectLockTestOneTSWithoutLease.TServerRefusesPGSessionsWithoutLeaseOnBoot'
./yb_build.sh fastdebug --cxx-test object_lock-test --gtest_filter 'ExternalObjectLockTestOneTS.TabletServerKillsSessionsWhenItAcquiresNewLease'

./yb_build.sh fastdebug --cxx-test process_wrapper-test
```

Reviewers: #db-approvers, amitanand, kramanathan, bkolagani

Reviewed By: amitanand

Subscribers: rthallam, slingam, ybase, yql, mlillibridge, svc_phabricator

Differential Revision: https://phorge.dev.yugabyte.com/D43586