[#26937] DocDB: YbHnsw persistence
Summary:
The YbHnsw library currently generates a block-based representation for the HNSW (Hierarchical Navigable Small World) graph structure. However, this representation is only stored in memory and lacks persistence, meaning it is lost when the application terminates or restarts.

To make YbHnsw a viable solution for implementing a vector index, we need to add functionality for saving the in-memory HNSW graph data to the file system and reloading it when needed. This persistence layer will ensure that the graph remains available across application sessions, improving usability and efficiency.

Also added basic wiring (not final interface) for block cache. To provide ability to load and unload blocks from the diff in followup diffs.
Jira: DB-16382

Test Plan: YbHnswTest.Persistence

Reviewers: arybochkin

Reviewed By: arybochkin

Subscribers: ybase

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D43473