[#26990] YSQL: Increase timeout for waiting for pg_yb_catalog_version mode
Summary:
In one of the QA stress tests, the following log was observed:

```
2025-04-28 14:45:14.891 UTC [34595] ERROR:  Failed to find out pg_yb_catalog_version mode: Operation 'catalog_version_table_in_perdb_mode is not set in shared memory' didn't complete within 10000ms
```

The gflag `--ysql_enable_db_catalog_version_mode` is used to enable per-database
catalog version mode. In reality, with YSQL upgrade, it is possible that
`--ysql_enable_db_catalog_version_mode=true` but the table
`pg_yb_catalog_version` has not been upgraded to have one-row-per-database yet.
In order to operate in per-database catalog version mode, both conditions must
be true.

The `pg_yb_catalog_version` mode refers to the second condition: if the table
has only one row (for tempalte1), it is in global catalog version mode; if the
table has one row per database, it is in per-database catalog version mode.

After a tserver starts (or restarts), it does not make a query of the table
`pg_yb_catalog_version` directly to find out its mode. Instead it finds out that
from the tserver/master heartbeat response because master will put the contents
of the table as part of the response. Before tserver finds out the mode, it is
unknown and it cannot begin servicing PG connections. Once it receives the
heartbeat response from master, it figures out the mode and stores the mode in
shared memory that can be accessed by all PG backends.

A PG backend, during its connection setup time, will read the shared memory for
this mode of `pg_yb_catalog_version`. If the mode is not set yet (unknown), it
cannot move forward. It keeps polling the shared memory for this mode until it
is set (either global or per-database mode), or until a 10 second timeout.

It was found that for the above test the master took 15 seconds to respond to
tserver's first heartbeat request, in fact the heartbeat request RPC itself has a 15
second timeout:

```
DEFINE_RUNTIME_int32(heartbeat_rpc_timeout_ms, 15000,
    "Timeout used for the TS->Master heartbeat RPCs.");
```

We should increase the timeout indicated in this log from 10000ms to more than
15 seconds. In the above failed test case 20 second timeout would have made the
case successful.

This diff increases the timeout from 10000ms to 20000ms, also added to the error
text to indicate tserver isn't ready for servicing queries yet.
Jira: DB-16465

Test Plan: /yb_build.sh debug --cxx-test pg_catalog_version-test --gtest_filter PgCatalogVersionTest.SimulateDelayedHeartbeatResponse

Reviewers: kfranz, sanketh, mihnea

Reviewed By: sanketh

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D43601