[#26526] Docdb: Table locks. Allow releasing exclusive/global locks through yb-admin
Summary:
Add a yb-admin command to release global/exclusive table locks through the master.

This is primarily a debug tool to clean up persisted locks at the master -- in case our existing mechanisms for cleaning up the locks are not enough and there is an unexpected leak.

For releasing shared/DML locks on TServer, a simpler approach may be to restart the TServer (if possible). However, for exclusive/global locks restart does not help as they are persisted.

The expected workflow would be for the user/engineer to curl the master's web-ui to identify the leaked locks and use the tool to release them -- ** after ** ensuring that the transaction is aborted (using `yb_cancel_transaction` as necessary).

To prevent accidental misuse, this command is hidden from the displayed options in yb-admin.

other minor changes:
- do not pass server_address for uuid in yb-ts-cli
- remove acquire codepath from yb-ts-cli

**Upgrade/Rollback safety:**
  This feature is guarded by the test flag `TEST_enable_object_locking_for_table_locks` . Unless the flag is enabled, these changes do not have any effect.
Jira: DB-15892

Test Plan: yb_build.sh fastdebug --cxx-test yb-admin-test --gtest_filter  AdminCliTestForTableLocks.TestYsqlTableLock

Reviewers: bkolagani, zdrudi, rthallam

Reviewed By: zdrudi

Subscribers: yql, ybase

Differential Revision: https://phorge.dev.yugabyte.com/D43615