[#27002] docdb: Fix PgGetLockStatusTest to handle uninvolved tablets
Summary:
**Issue:**
When `GetLockStatus` is called with specific txn ids, the tserver's response includes lock info from all tablets, including those not involved in the specified transactions(This is a known limitation tracked in issue #16913, which aims to make the response include only the involved tablets for request specifies txn ids). However, tests in PgGetLockStatusTest assume the response only contains involved tablets.
Those tests that call `GetLockStatus` with specific transaction ids are not failing today because luckly all tablets on the tserver happen to be involved.
The issue start becomes visible when advisory locks feature are enabled by default—since the advisory lock tablet will start to appear in the  GetLockStatus response even if it holds no locks.

**Fix:**
Update `VerifyResponse` and `GetNumTabletsInLockStatusResponse` to correctly handle the presence of uninvolved tablets in the response
Jira: DB-16476

Test Plan: Jenkins: test regex: .*PgGetLockStatusTest.*

Reviewers: bkolagani, rthallam

Reviewed By: bkolagani, rthallam

Subscribers: ybase, yql

Differential Revision: https://phorge.dev.yugabyte.com/D43621