[BACKPORT 2025.1][#26984] docdb: VectorLSMTest.CompactionCancelOnShutdown/kHnswlib failure
Summary:
HNSW uses SSE instructions widely to make the performance better. In some cases it skips boundaries
check in favor of performance. One of such scenarios in `searchBaseLayer` method lead to the attempt
accessing an memory past the allocated chunk, and as a result `_mm_prefetch` would try to read by
a random address, which is probably an undefined behaviour and may lead to a crash in theory:
```
_mm_prefetch((char *) (visited_array + *(datal + j + 1)), _MM_HINT_T0);
_mm_prefetch(getDataByInternalId(*(datal + j + 1)), _MM_HINT_T0);
```
where `j + 1` is the index of the item after the last available.

To prevent this behaviour and to follow the logic of the library, the fix extends the size of
the allocated memory by the size of the pointer, which covers all out-of-bound reads in SSE case.
Jira: DB-16457

Original commit: 1df539157e2c50b3ab75a3e3abbe31d416a9ea77 / D43622

Test Plan: ./yb_build.sh asan --cxx-test vector_index_vector_lsm-test --gtest_filter VectorLSMTest.CompactionCancelOnShutdown/kHnswlib

Reviewers: sergei, slingam

Reviewed By: slingam

Subscribers: ybase

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D43711