[PLAT-17181]: ysql_pg_conf_csv is not updating correctluy after a concurrent gflag upgrade
Summary:
Issue:
1. Try changing the pd_audit settings directly via the ‘ysql_pg_conf_csv’ concurrent gflag upgrade.
The issue happens when you try and edit an already set gflag that is enclosed in double quotes. UI seems to automatically remove the quotes in edit view and if the user is not careful enough and saves it, UI doesn’t add it back.

This behavior needs to be enforced on HBA CONF but not in PG CONF as PG CONF criteria for unformatting string is different
We have fixed the quotes unformatting issue for PG_CONF

2. We are showing red dot on top of Actions dropdown in OVERVIEW page. We show this as we have UPGRADE AVAILABLE. However if you open Action menu there you want’t see any message or a following dot in Upgrade Database version.
This does not make any sense to show this dot, we already show 'Upgrade available with up arrow button' in DB version widget. Hence removed the dot for now.

Test Plan:
Please refer to BEFORE and AFTER video

BEFORE fix
{F350959}

AFTER fix
{F350964}

Reviewers: jmak

Reviewed By: jmak

Differential Revision: https://phorge.dev.yugabyte.com/D43370