[#27015] docdb: Fix flaky PgCloneTest.CloneTimeoutExceeded
Summary:
The test was flaky because the command line uses the same timeout as the pg clone operation, so if the command line timed out before the actual clone, we could see the clone in the `CLONE_SCHEMA_STARTED` state instead of `ABORTED`. Fixed by changing to wait for the state to be `ABORTED`. Also added an artificial delay to the pg clone schema step in the unlikely event that step runs faster than the specified timeout.
Jira: DB-16489

Test Plan: Set the client timeout to 1ms and reproduced the behavior, which is fixed after my changes.

Reviewers: mhaddad

Reviewed By: mhaddad

Subscribers: ybase

Differential Revision: https://phorge.dev.yugabyte.com/D43639