[#26891] CDC: Fix schema packing not found issue with index creation
Summary:
When CDC is active and a user creates an index, there is a possibility that cdc producer hits `Schema packing not found` issue. This issue can arise when WAL has out of order apply of transactions interleaved with CHANGE_METADATA_OPs from index creation. Once this error is hit, the stream / slot is rendered useless and the only way to proceed is to create a new stream / slot.

In `CDCServiceImpl` we maintain a `schema_details_map` which maintains a mapping between the table and its schema relevant to CDC. Note that since CDC could be lagging this schema may not always be the latest schema for the table. To decode packed rows a `SchemaPackingStorage` is required. `SchemaPackingStorage` maintains a mapping of schema_version and schema_packing (derivative of Schema). In CDC producer, for each batch of intents (or WRITE_OP) we create a new `SchemaPackingStorage` and initialize it with the schema present in  `schema_details_map`. Because of this, although `SchemaPackingStorage` is capable of storing schemas corresponding to multiple schema versions, we end up storing just one schema.

During Index creation, `schema_details_map` will be updated when a CHANGE_METADATA_OP is encountered. Now, if we receive any intents which require previous schema version for decoding after the CHANGE_METADATA_OP, we will hit `Schema packing not found` issue. To fix this 2 things are done in this revision:

  # Maintain a map from table_id to `SchemaPackingStorage` in `CDCServiceImpl`. This increases the scope of `SchemaPackingStorage` to class level, i.e now we will have entries corresponding to multiple schema versions stored in `SchemaPackingStorage`. Currently the scope for `SchemaPackingStorage` is only a single GetChanges call. After this revision the scope of SchemaPackingStorage will be the same as CDC service. This map (called `table_to_schema_packing_storage_`) will be guarded by the same mutex which guards `cdc_state_metadata_` (which contains `schema_details_map`) and `tablet_checkpoints_` map. In each GetChanges call, a copy of `table_to_schema_packing_storage_` will be sent to CDC producer. CDC producer will add entries to this copy if required and at the end of GetChanges, `table_to_schema_packing_storage_` will be updated with the modified copy. This mechanism is similar to what is followed for `schema_details_map`.  The newly added `table_to_schema_packing_storage_` will be cleared only at the shut down of CDC service. There is no mechanism implemented in this revision to remove entries from `table_to_schema_packing_storage_`. However, this should be fine as we do not expect too many schema version changes when CDC is active.

  # When CDC producer does not find an entry in `schema_details_map` corresponding to the table whose intents it is decoding, it fetches the schema from master (`GetTableSchemaFromSysCatalog` rpc) and populates the map. `GetTableSchemaFromSysCatalog` rpc takes read_time as a parameter in its request. Currently we send the hybrid time of the first intent in the batch as the read_time. This could be buggy if a batch of intents contain intents requiring different schema_versions for decoding. This revision fixes this by sending the hybrid time of the intent for which appropriate schema version was not found in `SchemaPackingStorages`.

Other changes:

  - This diff changes the default value of the test flag `dcheck_for_missing_schema_packing` to false. This flag when true creates a FATAL in debug builds when `SchemaPackingStorage::GetPacking()` is not able to find required schema version. We wanted to use this function in CDC producer to detect the missing schema and populate it accordingly.
  - We used to populate a DDL record in CDC producer even when there are no other records to stream. This behaviour is now fixed as a side effect of these changes. As a result the assertion in `CDCSDKYsqlTest.TestUpdateOnNonExistingEntry` has been changed accordingly.
Jira: DB-16301

Test Plan:
Jenkins: urgent
A connector side UT will be added in a separate PR which reproduces this issue.
Stress test where the issue was originally found was run against a custom build.

Reviewers: skumar, stiwary, aagrawal, vkushwaha

Reviewed By: stiwary

Subscribers: ycdcxcluster, skumar

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D43537