[#26751] YSQL: Move Cancel method from ScopeExit into CancellableScopeExit class
Summary:
The `ScopeExitLambda::Cancel` method was introduced in the https://phorge.dev.yugabyte.com/D31684 diff.
This method allows to dismiss execution of a lambda on exit from scope. This feature is useful in some very specific situations and it is used very rarely. Nowadays YB has 4 occurrences of such a code which can be reduced to 2. While the usage of ScopeExitLambda itself quite huge (178 occurrences in total). So the cancelability is used in about 1% of cases.
But having the `Cancel` method in `ScopeExitLambda` class makes all the usage of `ScopeExitLambda` potentially cancelable. While it is definitely not intended by design. Also this limits the usage of `ScopeExitLambda` in the interfaces as the guard object. Like this one

```
auto DisabledReadTimeHistoryUpdate() {
  const auto old_value = lock_;
  lock_ = true;
  return ScopedExit([&lock = lock_, old_value] { lock = old_value; });
}
```

Because there is no guaranty that the caller code will not call `Cancel` method on a guard object. As a result developer is forced to create hand made class for his/her needs which increases the code base.

Based on the above it is reasonable to create separate `CancelableScopeExit` class for situations when cancelability is really necessary and make `ScopeExit` class non cancelable.

Major changes:
- The `ScopeExit` helper function is completely removed (it is redundant due to `CTAD`)
- `ScopedExitLambda` template is renamed into `CancellableScopeExit`
- New template class `ScopeExit` is introduced which encapsulate `CancellableScopeExit` and hides `Cancel` method

Additional changes:
- Remove several redundant usage of cancellable scoped exit lambda
Jira: DB-16135

Test Plan: Jenkins

Reviewers: arybochkin, hsunder, zdrudi, mlillibridge, esheng, sergei, #db-approvers

Reviewed By: arybochkin, hsunder, mlillibridge, sergei

Subscribers: yql

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D43324