[#27052] CDC: Add test for colocated + index issue
Summary:
When streaming data with colocation, cdcsdk_producer reads intents for all the tables (including the index tables) residing on the colocated tablet. We skip shipping records for the tables which are not present in stream metadata. However, due to erroneous handling of schema used for decoding, we would end up updating the required schema with the schema of an index table. This would result in skipping of intents while decoding, which manifested as loss of few records or loss of columns on which index was created.

The issue has been fixed by D43537. This revision adds a UT for this case.
Jira: DB-16526

Test Plan:
Jenkins: test regex: .*CDC.*
./yb_build.sh --cxx-test integration-tests_cdcsdk_consumption_consistent_changes-test --gtest_filter CDCSDKConsumptionConsistentChangesTest.TestColocatedUpdateWithIndex

Reviewers: vkushwaha, skumar

Reviewed By: skumar

Subscribers: ycdcxcluster

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D43706