[PLAT-17520] Introduce a field in universe details that is updated whenever node agents are uninstalled or installed on any universe node
Summary:
Periodically update the universe details field to know if there are missing node agents for a universe. This is needed to serve the UI banner to warn customers. The update is not done if the universe lock is already held by some
other tasks.

Test Plan: UT + manually tested

Reviewers: amalyshev, jmak, nbhatia

Reviewed By: jmak

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D43732