[#27073] YSQL: Fix org.yb.pgsql.TestPgConnection#testConnectionKills with Connection Manager
Summary:
This diff fixes `org.yb.pgsql.TestPgConnection#testConnectionKills` test with connection manager enabled. The test has already been disabled to be run with connection manager but still fails on running with conn mgr.
The issue was before entering to test or changing the port to use PgPort (on skipping the test with connection manager), there are some connections already been created to conn mgr port as part of initial setup. Same connection is used in cleanUpAfter function which executes the query on the conn mgr port by creating additional transactional backend process.
Therefore fix is to have a clean state before starting the test by restarting the cluster, so all connections are created only on pgPort and no interference of connections to conn mgr port is there when test is made to skip with conn mgr.
Jira: DB-16562

Test Plan: Jenkins: enable connection manager, .*TestPgConnection.*

Reviewers: rbarigidad, skumar

Reviewed By: skumar

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D43693