[PLAT-17072] Aggregate slow queries data in system_platform DB
Summary:
This diff adds a scheduled task which runs once an hour and stores slow queries data.
The data is stored in `system_platform` DB of each universe.
We delete the data after 7 days and this is runtime configurable.
This should help us compare query statistics before/after a given timestamp.

Test Plan: Manually verified that the table is created for a universe and data is correctly populated.

Reviewers: amalyshev, vbansal

Reviewed By: amalyshev

Differential Revision: https://phorge.dev.yugabyte.com/D43789