[PLAT-17555][YBA CLI]Make storage-config as optional so that xcluster can be created without bootstrap
Summary: Some customers would want to skip full copy of tables while creating/editing replication. Storage configurations are only required when tables need full copy. Setting the flag as optional and introducing `skip-full-copy-tables` to fix the same

Test Plan:
Create 2 universes A (source) and B (target) and create tables with the same name in both
Add entries in the table in universe A
Create xCluster between A and B with `--skip-full-copy-tables` enabled
Add more entries in the table in universe A

Verify that table in universe B contains only the entries made in table from A after replication is enabled

Reviewers: hzare

Reviewed By: hzare

Differential Revision: https://phorge.dev.yugabyte.com/D43841