[BACKPORT 2025.1][#26951] YSQL: Fix RETURNING ybctid in DML queries
Summary:
For DML queries, the values in the `tupleSlot` and `planSlot` are used to compute RETURNING projections.
This implies that if the ybctid system column is used in the RETURNING clause, it must be populated in the appropriate slot.
More specifically,
 - For INSERTs and UPDATEs, the `tupleSlot` contains the NEW row and must populate the ybctid.
 - For DELETEs, the `planSlot` contains the OLD row, as there is no NEW row, and must populate the ybctid.

This revision ensures that the appropriate slots always populate the ybctid for DML operations.
Jira: DB-16401

Original commit: 1ac5a3b90fd7d25615bfb4020a24bd86600f78c3 / D43616

Test Plan:
Run the following test:
```
./yb_build.sh --java-test 'org.yb.pgsql.TestPgRegressYbctid'
```

Reviewers: #db-approvers, aagrawal, amartsinchyk, smishra

Reviewed By: aagrawal

Subscribers: yql, smishra, svc_phabricator

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D43868