[#27114] YSQL: Error out of ANALYZE in mixed mode
Summary:
ANALYZE command can't work in the mixed PG11 to PG15 mode for two main
reasons:

1. The random number generator has changed between PG11 and PG15 and its
   state went from 48 to 128 bits. There's no way to convert one into
   other. In our implementation, sinse these states are the different
   protobuf fields, when sampling state is passed between different base
   Postgres versions the receiving side reads uninitialized random
   number generator state.
2. We prohibit metadata updates in mixed mode, so ANALYZE can't record
   collected statistics, even if it succeeds.

Hence we do not try to make ANALYZE command working, but error out early
and gracefully when mixed mode is detected. In this diff we are adding a
validation, whether the PG15 random number generator state presents in
the incoming sampling state data. If not, that indicates that the state
is coming from incompatible node. We add this validation in two places,
in DocDB when it receives the request from PgGate, and in PgGate where
it receives response from DocDB.

It is hard to add equivalent validation to PG11. We have no control,
from what PG11 based version user upgrades to PG15. It may happen that
the version is too old and has no validation. Apparently, it is not a
problem. If PG11 PgGate sends request to PG15 DocDB, it returns an error.
If PG15 PgGate sends request to PG11 DocDB it proceeds with
uninitialized random number generator state, however it is resilient to
that. It collects potentially incorrect portion of the sample, but the
response is rejected by PG15, so wasted work on PG11 node is the only
impact.

**Upgrade/Rollback safety:**
Comment only change to a .proto file
Jira: DB-16600

Test Plan:
Manual.
Start from a PG11 based cluster.
Start upgrade.
 - ANALYZE fails because it can't update statistics
Upgrade some tservers
 - ANALYZE fails because of missing random state both when ysqlsh is connected to PG11 or PG15 node.
Upgrade remaining tservers
 - ANALYZE fails because it can't update statistics
Finalize upgrade
 - ANALYZE succeeds

Unit:
./yb_build.sh --gcc13 --cxx-test ysql_major_upgrade-test --gtest_filter YsqlMajorUpgradeTest.Analyze

Reviewers: aagrawal, telgersma, timur

Reviewed By: telgersma

Subscribers: yql

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D43863