[PLAT-17437] Add enable node agent migration flow to YBA UI
Summary:
This diff adds a banner to the dashboard page which appears whenever
the user does not have the runtime config `yb.node_agent.enabler.run_installer` set
to `true`.
This banner serves to nudge the user into enabling this runtime config flag so YBA is able
to automatically install Node Agent on all existing universes without Node Agent.

Here is the related backend change whcih introduced the runtime config flag for
automatically installing Node Agent:
e872607bacf8895b6b966507fc4d6228bfdd72ad / D43566

The change was introduced to let users opt-in to the migration at a time which is
convenient for them instead of pushing the migration onto users as soon as they upgrade their YBA
version.

It is worth noting that the `nodeAgentMissing` field does not get updated if a universe is paused.

Test Plan:
Verify the Node Agent banner text.
Verify the `Enable Automatic Node Agent Migration` button opens a modal for the user to confirm
their intent to enable the migration.
Verify the modal appearance is as expected.
Verify the modal submission modifies the runtime config.
{F353917}

**Case 1: Automatic Node Agent migration disabled & no installation failures**
{F355715}

**Case 2: Automatic Node Agent migration disabled & has installation failures**
{F355714}

**Case 3: Automatic Node Agent migration enabled & no installation failures**
{F355712}

**Case 4: Automatic Node Agent migration enabled & has installation failures**
{F355713}

Verify that `nodeAgentMissing` gets set to true if the user manually remove node agent from a universe
node. If a universe is paused, the field does not get updated until the universe is resumed.

Reviewers: nsingh, rmadhavan, nbhatia

Reviewed By: nsingh

Differential Revision: https://phorge.dev.yugabyte.com/D43602